# api/openapi/apiserver/v1/rca-complete.swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RCA API Complete",
    "version": "v1",
    "description": "Complete API surface scanned from Gin route registrations with RBAC hints."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/v1/config/pipeline/{alert_source}": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "GetPipelineConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `GetPipelineConfig`.",
        "operationId": "GetPipelineConfig_GET_v1_config_pipeline_alert_source",
        "parameters": [
          {
            "name": "alert_source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "GetPipelineConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go"
      }
    },
    "/v1/config/pipeline/update": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "UpsertPipelineConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `UpsertPipelineConfig`.",
        "operationId": "UpsertPipelineConfig_POST_v1_config_pipeline_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "UpsertPipelineConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "alert_source": {
                    "type": "string"
                  },
                  "service": {
                    "type": "string"
                  },
                  "namespace": {
                    "type": "string"
                  },
                  "pipeline_id": {
                    "type": "string"
                  },
                  "graph_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/config/sla/{session_type}": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "GetSLAConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `GetSLAConfig`.",
        "operationId": "GetSLAConfig_GET_v1_config_sla_session_type",
        "parameters": [
          {
            "name": "session_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "GetSLAConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go"
      }
    },
    "/v1/config/sla/update": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "UpsertSLAConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `UpsertSLAConfig`.",
        "operationId": "UpsertSLAConfig_POST_v1_config_sla_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "UpsertSLAConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "session_type": {
                    "type": "string"
                  },
                  "due_seconds": {
                    "type": "integer"
                  },
                  "escalation_thresholds": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/config/toolset/{pipeline_id}": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "GetToolsetConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `GetToolsetConfig`.",
        "operationId": "GetToolsetConfig_GET_v1_config_toolset_pipeline_id",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "GetToolsetConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go"
      }
    },
    "/v1/config/toolset/update": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "UpsertToolsetConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `UpsertToolsetConfig`.",
        "operationId": "UpsertToolsetConfig_POST_v1_config_toolset_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "UpsertToolsetConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pipeline_id": {
                    "type": "string"
                  },
                  "toolset_name": {
                    "type": "string"
                  },
                  "allowed_tools": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/config/trigger/{trigger_type}": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "GetTriggerConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `GetTriggerConfig`.",
        "operationId": "GetTriggerConfig_GET_v1_config_trigger_trigger_type",
        "parameters": [
          {
            "name": "trigger_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "GetTriggerConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go"
      }
    },
    "/v1/config/trigger/update": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "UpsertTriggerConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `config.admin`. Handler: `UpsertTriggerConfig`.",
        "operationId": "UpsertTriggerConfig_POST_v1_config_trigger_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "config.admin",
        "x-handler": "UpsertTriggerConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trigger_type": {
                    "type": "string"
                  },
                  "pipeline_id": {
                    "type": "string"
                  },
                  "session_type": {
                    "type": "string"
                  },
                  "fallback": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/session/{sessionID}/assign": {
      "post": {
        "tags": [
          "SessionConfig"
        ],
        "summary": "AssignSessionConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `session.assignment`. Handler: `AssignSessionConfig`.",
        "operationId": "AssignSessionConfig_POST_v1_session_sessionID_assign",
        "parameters": [
          {
            "name": "sessionID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "session.assignment",
        "x-handler": "AssignSessionConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "assignee"
                ],
                "properties": {
                  "assignee": {
                    "type": "string"
                  },
                  "assigned_by": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/session/{sessionID}/assignment": {
      "get": {
        "tags": [
          "SessionConfig"
        ],
        "summary": "GetSessionAssignmentConfig",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `ai.read|session.assignment`. Handler: `GetSessionAssignmentConfig`.",
        "operationId": "GetSessionAssignmentConfig_GET_v1_session_sessionID_assignment",
        "parameters": [
          {
            "name": "sessionID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "ai.read|session.assignment",
        "x-handler": "GetSessionAssignmentConfig",
        "x-source-file": "internal/apiserver/handler/internal_strategy_config.go"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "reason": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  }
}